Upgrade to Scala 2.13.7
Mostly handled by Claude, with a bunch of steering.

At this point, everything compiles. Haven't done tests or tried running for real yet.

*Most* of this is simple stuff -- in particular, getting rid of procedure syntax, cleaning up arity-zero vs empty parameter list, and cleaning up unused imports. A few changes are slightly more significant, but nothing radical -- I told it to keep the changes conservative.

A lot of deprecation warnings have been suppressed for now (by Claude): need to go back and fix those, one by one.

+++ querki/scalajs/src/main/scala/querki/time/package.scala
@@ -11,7 +11,7 @@ package object time {
   // HACK to get common time formatting. In the medium term this needs to become much smarter,
   // to respond correctly to localization.
   def displayTime(time: Common.Timestamp): String = {
-    Moment(time).calendar(
+    Moment(time.toDouble).calendar(
       js.undefined,
       lit(
         sameElse = "MM/DD/YYYY"
